requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
price-parser==0.3.4
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
yarl==1.18.3
